Add sub-table ranking capability (ugly, but functional)
The algorithm is mainly in class Group, methods printTable,
refreshTable and integrateSubTable. Class TableRow with methods
compareTo and updateMatchStatistics is also instrumental.

TODO: All in all this is an ugly, quick & dirty solution, a mere proof
of concept. It needs heavy refactoring to make it more readable, robust
and aesthetic in general. :-/ Ranking table printing is also just
rudimentary, using TableRow.toString.