Refactoring: introduce UpdateableTreeSet
The new collection class in combination with its public inner interface
UpdateableTreeSet.Updateable enables updates on tree sets while looping
over them. More exactly, the elements are marked for update. The update
as such is deferred until later (after the loop has finished) and then
done as a bulk operation. This helps us get rid of 'copyToArray'
constructs, i.e. explicit copies of collections just for the sake of
looping over them safely, updating the original collection on the fly.

Implementation detail: UpdateableTreeSet keeps internal lists of
elements to be removed or updated - well, not really lists but
IdentityHashMaps. So now the loops are hidden inside the collection
class, not scattered throughout the user code anymore.