Added Primera División 2006/07
This is an especially interesting example because in that year the
championship was decided in favour of Real against Barca by direct
comparison at equal points, even though Barca's overall GD (goal
difference) was significantly better than Real's. But direct comparison
speaks a clear language of 4-1 points in favour of Real.