Table.Row: new member goalsAway (goals scored away from home)
See http://en.wikipedia.org/wiki/Away_goal. The new member is now also
printed as a table column with heading "GW" (because "GA" was already
taken by member goalsAgainst). Currently thie member is irrelevant for
the ranking, but it might be of importance when the program is extended
to calculate other types of tables or e.g. UEFA Champions League results
during elimination (knock-out) phase.