New static factory method TableRowComparator.fromStrategy(String)
The new method can construct a nested hierarchy of comparators by
parsing a String like "Foo(Bar, Zot(One, Two)), Baz, Bla". Therein a
comma designates a successor, an pair of brackets designate a child.
Structures can be nested like in the example above because the factory
method is used recursively.

The factory method expects all concrete comparators to be constructed to
live in the same package as TableRowConstructor, to have an empty
default constructor and to have a class name ending with "Comparator",
i.e. a short name like "Foo" becomes my.package.FooComparator.

Class Ranking was refactored to use the new factory method, so please
see there for usage examples.