Refactor & improve sub-table printing: no more doubles in output
Sub-tables are now not just dumped as they are created, but collected
as Strings (not as Table objects) in an internal LinkedHashSet,
maintaining the chronological order in which they were added to the set.
This avoids a lot of doubles in the output caused by sub-tables created
for each call to Table.Row.compare during Collections.sort.

In order to avoid another type of doublet caused by mathematically
identical sub-table rows in random order, a new Comparator subclass
named TeamIdComparator was introduced. It acts as a wrapper/decorator
around the usual TableRowComparators and takes care of stable
alphabetical ordering by team ID in case of otherwise identical ranking.

Another improvement is that now hierarchical indentation of sub-tables
is supported for nested parent-child relationships. Until now all levels
of sub-tables were printed with the same indentation. Actually there are
no such cases in any ranking systems I know of, but I have implemented
and tested it anyway, just in case. :-)